LoopBuilder: extract code for binding an action to samplers
