+++ src/main/java/net/imglib2/loops/LoopBuilder.java
@@ -47,7 +47,6 @@ import net.imglib2.Interval;
 import net.imglib2.Positionable;
 import net.imglib2.RandomAccess;
 import net.imglib2.RandomAccessibleInterval;
-import net.imglib2.Sampler;
 import net.imglib2.util.Intervals;
 
 /**
@@ -130,12 +129,12 @@ public class LoopBuilder< T >
 	public void forEachPixel( final T action )
 	{
 		Objects.requireNonNull( action );
-		final List< RandomAccess< ? > > samplers = Stream.of( images ).map( this::initRandomAccess ).collect( Collectors.toList() );
+		final List< RandomAccess< ? > > samplers = Stream.of( images ).map( LoopBuilder::initRandomAccess ).collect( Collectors.toList() );
 		final Positionable synced = SyncedPositionables.create( samplers );
-		LoopUtils.createIntervalLoop( synced, dimensions, RunnableFactory.bindActionToSamplers( action, samplers ) ).run();
+		LoopUtils.createIntervalLoop( synced, dimensions, BindActionToSamplers.bindActionToSamplers( action, samplers ) ).run();
 	}
 
-	private RandomAccess< ? > initRandomAccess( final RandomAccessibleInterval< ? > image )
+	private static RandomAccess< ? > initRandomAccess( final RandomAccessibleInterval< ? > image )
 	{
 		final RandomAccess< ? > ra = image.randomAccess();
 		ra.setPosition( Intervals.minAsLongArray( image ) );
@@ -160,230 +159,5 @@ public class LoopBuilder< T >
 	public interface SixConsumer< A, B, C, D, E, F >
 	{
 		void accept( A a, B b, C c, D d, E e, F f );
-	}
-
-	private static class RunnableFactory
-	{
-
-		private static final List< ClassCopyProvider< Runnable > > factories = Arrays.asList(
-				new ClassCopyProvider<>( ConsumerRunnable.class, Runnable.class ),
-				new ClassCopyProvider<>( BiConsumerRunnable.class, Runnable.class ),
-				new ClassCopyProvider<>( TriConsumerRunnable.class, Runnable.class ),
-				new ClassCopyProvider<>( FourConsumerRunnable.class, Runnable.class ),
-				new ClassCopyProvider<>( FiveConsumerRunnable.class, Runnable.class ),
-				new ClassCopyProvider<>( SixConsumerRunnable.class, Runnable.class ));
-
-		/**
-		 * For example.: Given a BiConsumer and two Samplers:
-		 *
-		 * <pre>
-		 * {@code
-		 * BiConsumer<A, B> biConsumer = ... ;
-		 * Sampler<A> samplerA = ... ;
-		 * Sampler<B> samplerB = ... ;
-		 * }
-		 * </pre>
-		 *
-		 * This method
-		 * {@code bindConsumerToSamplers(biConsumer, Arrays.asList(samplerA, samplerB))}
-		 * will return a Runnable that is functionally equivalent to:
-		 *
-		 * <pre>
-		 * {
-		 * 	&#64;code
-		 * 	Runnable result = () -> {
-		 * 		biConsumer.accept( sampleA.get(), samplerB.get() );
-		 * 	};
-		 * }
-		 * </pre>
-		 *
-		 * It does it in such manner, that the returned {@link Runnable} can be
-		 * gracefully optimised by the Java just-in-time compiler.
-		 *
-		 * @param action
-		 *            This must be an instance of {@link Consumer},
-		 *            {@link BiConsumer} of {@link TriConsumer}.
-		 * @param samplers
-		 *            A list of {@link Sampler}, the size of the list must fit
-		 *            the consumer given by {@param operation}.
-		 * @throws IllegalArgumentException
-		 *             if the number of sampler does not fit the given consumer.
-		 */
-		public static Runnable bindActionToSamplers( final Object action, final List< ? extends Sampler< ? > > samplers )
-		{
-			final Object[] arguments = Stream.concat( Stream.of( action ), samplers.stream() ).toArray();
-			for ( final ClassCopyProvider< Runnable > factory : factories )
-				if ( factory.matches( arguments ) )
-				{
-					final List< Class< ? extends Object > > key = Stream.of( arguments ).map( Object::getClass ).collect( Collectors.toList() );
-					return factory.newInstanceForKey( key, arguments );
-				}
-			throw new IllegalArgumentException();
-		}
-
-		public static class ConsumerRunnable< A > implements Runnable
-		{
-
-			private final Consumer< A > action;
-
-			private final Sampler< A > samplerA;
-
-			public ConsumerRunnable( final Consumer< A > action, final Sampler< A > samplerA )
-			{
-				this.action = action;
-				this.samplerA = samplerA;
-			}
-
-			@Override
-			public void run()
-			{
-				action.accept( samplerA.get() );
-			}
-		}
-
-		public static class BiConsumerRunnable< A, B > implements Runnable
-		{
-
-			private final BiConsumer< A, B > action;
-
-			private final Sampler< A > samplerA;
-
-			private final Sampler< B > samplerB;
-
-			public BiConsumerRunnable( final BiConsumer< A, B > action, final Sampler< A > samplerA, final Sampler< B > samplerB )
-			{
-				this.action = action;
-				this.samplerA = samplerA;
-				this.samplerB = samplerB;
-			}
-
-			@Override
-			public void run()
-			{
-				action.accept( samplerA.get(), samplerB.get() );
-			}
-		}
-
-		public static class TriConsumerRunnable< A, B, C > implements Runnable
-		{
-
-			private final TriConsumer< A, B, C > action;
-
-			private final Sampler< A > samplerA;
-
-			private final Sampler< B > samplerB;
-
-			private final Sampler< C > samplerC;
-
-			public TriConsumerRunnable( final TriConsumer< A, B, C > action, final Sampler< A > samplerA, final Sampler< B > samplerB, final Sampler< C > samplerC )
-			{
-				this.action = action;
-				this.samplerA = samplerA;
-				this.samplerB = samplerB;
-				this.samplerC = samplerC;
-			}
-
-			@Override
-			public void run()
-			{
-				action.accept( samplerA.get(), samplerB.get(), samplerC.get() );
-			}
-		}
-
-		public static class FourConsumerRunnable< A, B, C, D > implements Runnable
-		{
-
-			private final FourConsumer< A, B, C, D > action;
-
-			private final Sampler< A > samplerA;
-
-			private final Sampler< B > samplerB;
-
-			private final Sampler< C > samplerC;
-
-			private final Sampler< D > samplerD;
-
-			public FourConsumerRunnable( final FourConsumer< A, B, C, D > action, final Sampler< A > samplerA, final Sampler< B > samplerB, final Sampler< C > samplerC, final Sampler< D > samplerD )
-			{
-				this.action = action;
-				this.samplerA = samplerA;
-				this.samplerB = samplerB;
-				this.samplerC = samplerC;
-				this.samplerD = samplerD;
-			}
-
-			@Override
-			public void run()
-			{
-				action.accept( samplerA.get(), samplerB.get(), samplerC.get(), samplerD.get() );
-			}
-		}
-
-		public static class FiveConsumerRunnable< A, B, C, D, E > implements Runnable
-		{
-
-			private final FiveConsumer< A, B, C, D, E > action;
-
-			private final Sampler< A > samplerA;
-
-			private final Sampler< B > samplerB;
-
-			private final Sampler< C > samplerC;
-
-			private final Sampler< D > samplerD;
-
-			private final Sampler< E > samplerE;
-
-			public FiveConsumerRunnable( final FiveConsumer< A, B, C, D, E > action, final Sampler< A > samplerA, final Sampler< B > samplerB, final Sampler< C > samplerC, final Sampler< D > samplerD, Sampler< E > samplerE )
-			{
-				this.action = action;
-				this.samplerA = samplerA;
-				this.samplerB = samplerB;
-				this.samplerC = samplerC;
-				this.samplerD = samplerD;
-				this.samplerE = samplerE;
-			}
-
-			@Override
-			public void run()
-			{
-				action.accept( samplerA.get(), samplerB.get(), samplerC.get(), samplerD.get(), samplerE.get() );
-			}
-		}
-
-		public static class SixConsumerRunnable< A, B, C, D, E, F > implements Runnable
-		{
-
-			private final SixConsumer< A, B, C, D, E, F > action;
-
-			private final Sampler< A > samplerA;
-
-			private final Sampler< B > samplerB;
-
-			private final Sampler< C > samplerC;
-
-			private final Sampler< D > samplerD;
-
-			private final Sampler< E > samplerE;
-
-			private final Sampler< F > samplerF;
-
-			public SixConsumerRunnable( final SixConsumer< A, B, C, D, E, F > action, final Sampler< A > samplerA, final Sampler< B > samplerB, final Sampler< C > samplerC, final Sampler< D > samplerD, final Sampler< E > samplerE, final Sampler< F > samplerF )
-			{
-				this.action = action;
-				this.samplerA = samplerA;
-				this.samplerB = samplerB;
-				this.samplerC = samplerC;
-				this.samplerD = samplerD;
-				this.samplerE = samplerE;
-				this.samplerF = samplerF;
-			}
-
-			@Override
-			public void run()
-			{
-				action.accept( samplerA.get(), samplerB.get(), samplerC.get(), samplerD.get(), samplerE.get(), samplerF.get() );
-			}
-		}
 	}
 }
